refactor: 프론트엔드 prettier 적용
* 코드 스타일 일관성을 위해 prettier 적용

+++ frontend/src/views/AdminView.vue
@@ -7,7 +7,7 @@
             <v-toolbar-title>유저 목록</v-toolbar-title>
           </v-toolbar>
         </template>
-        <template v-slot:item.actions="{ item }" >
+        <template v-slot:item.actions="{ item }">
           <v-icon small @click="check(item)">
             mdi-delete
           </v-icon>
@@ -29,30 +29,34 @@ export default {
         { text: "닉네임", align: "start", value: "nickname", sortable: false },
         { text: "전화번호", align: "start", value: "phone", sortable: false },
         { text: "성별", align: "start", value: "sex", sortable: false },
-        { text: "생년월일", align: "start", value: "birthday", sortable: false },
+        {
+          text: "생년월일",
+          align: "start",
+          value: "birthday",
+          sortable: false,
+        },
         { text: "Delete", align: "end", value: "actions", sortable: false },
       ],
       userlist: [],
       idx: 0,
-      email:'',
+      email: "",
     };
   },
   async created() {
     const { data } = await getallUser();
-    
+
     this.userlist = data;
   },
 
   methods: {
-    check(item){
-      confirm("해당 유저를 삭제하시겠습니까") &&
-        deleteUser(item);
+    check(item) {
+      confirm("해당 유저를 삭제하시겠습니까") && deleteUser(item);
     },
 
     async deleteUser(item) {
       const index = this.userlist.indexOf(item);
 
-      const Todelete={ email:this.userlist[index].email };
+      const Todelete = { email: this.userlist[index].email };
 
       try {
         const { data } = await deleteUser(Todelete);
@@ -62,7 +66,7 @@ export default {
         console.log(error);
       }
     },
-  }
+  },
 };
 </script>
